Only show errors when manually saving searches
This supresses errors when searches are saved for search history.
